move code to explorer.py
